// src/interfaces/country.interface.ts
interface NativeName {
  eng?: Eng;
  ara?: Ara;
  fra?: Fra;
  glv?: Glv;
  hat?: Hat;
  est?: Est;
  bul?: Bul;
  por?: Por;
  fas?: Fas;
  lav?: Lav;
  urd?: Urd;
  hin?: Hin;
  tam?: Tam;
  spa?: Spa;
  tur?: Tur;
  rus?: Rus;
  tgk?: Tgk;
  nya?: Nya;
  prs?: Prs;
  pus?: Pus;
  tuk?: Tuk;
  pih?: Pih;
  kor?: Kor;
  ita?: Ita;
  div?: Div;
  aym?: Aym;
  grn?: Grn;
  que?: Que;
  deu?: Deu;
  kir?: Kir;
  nfr?: Nfr;
  ind?: Ind;
  niu?: Niu;
  sqi?: Sqi;
  fij?: Fij;
  hif?: Hif;
  ton?: Ton;
  tsn?: Tsn;
  pau?: Pau;
  bis?: Bis;
  srp?: Srp;
  slk?: Slk;
  jpn?: Jpn;
  tir?: Tir;
  mri?: Mri;
  nzs?: Nzs;
  swa?: Swa;
  ces?: Ces;
  nrf?: Nrf;
  cha?: Cha;
  swe?: Swe;
  uzb?: Uzb;
  crs?: Crs;
  nld?: Nld;
  lao?: Lao;
  ell?: Ell;
  zho?: Zho;
  sot?: Sot;
  hmo?: Hmo;
  tpi?: Tpi;
  aze?: Aze;
  msa?: Msa;
  mah?: Mah;
  ron?: Ron;
  som?: Som;
  gil?: Gil;
  nep?: Nep;
  vie?: Vie;
  tvl?: Tvl;
  gsw?: Gsw;
  roh?: Roh;
  mlg?: Mlg;
  cat?: Cat;
  jam?: Jam;
  ben?: Ben;
  nno?: Nno;
  nob?: Nob;
  smi?: Smi;
  pap?: Pap;
  zdj?: Zdj;
  bos?: Bos;
  hrv?: Hrv;
  amh?: Amh;
  bar?: Bar;
  fin?: Fin;
  kon?: Kon;
  lin?: Lin;
  tet?: Tet;
  ber?: Ber;
  mey?: Mey;
  mon?: Mon;
  fil?: Fil;
  lua?: Lua;
  lit?: Lit;
  nor?: Nor;
  gle?: Gle;
  mya?: Mya;
  afr?: Afr;
  her?: Her;
  hgm?: Hgm;
  kwn?: Kwn;
  loz?: Loz;
  ndo?: Ndo;
  bel?: Bel;
  lat?: Lat;
  bjz?: Bjz;
  bwg?: Bwg;
  kck?: Kck;
  khi?: Khi;
  ndc?: Ndc;
  nde?: Nde;
  sna?: Sna;
  toi?: Toi;
  tso?: Tso;
  ven?: Ven;
  xho?: Xho;
  zib?: Zib;
  smo?: Smo;
  mlt?: Mlt;
  kat?: Kat;
  mfe?: Mfe;
  ltz?: Ltz;
  hye?: Hye;
  sag?: Sag;
  tha?: Tha;
  slv?: Slv;
  pol?: Pol;
  mkd?: Mkd;
  arc?: Arc;
  ckb?: Ckb;
  dan?: Dan;
  fao?: Fao;
  dzo?: Dzo;
  heb?: Heb;
  sin?: Sin;
  nau?: Nau;
  kal?: Kal;
  hun?: Hun;
  ssw?: Ssw;
  isl?: Isl;
  pov?: Pov;
  rar?: Rar;
  cnr?: Cnr;
  tkl?: Tkl;
  khm?: Khm;
  kaz?: Kaz;
  cal?: Cal;
  nbl?: Nbl;
  nso?: Nso;
  zul?: Zul;
  kin?: Kin;
  run?: Run;
  ukr?: Ukr;
}

interface Name {
  common: string;
  official: string;
  nativeName?: NativeName;
}

interface Ttd {
  name: string;
  symbol: string;
}

interface Sar {
  name: string;
  symbol: string;
}

interface Gyd {
  name: string;
  symbol: string;
}

interface Xaf {
  name: string;
  symbol: string;
}

interface Gbp {
  name: string;
  symbol: string;
}

interface Imp {
  name: string;
  symbol: string;
}

interface Shp {
  name: string;
  symbol: string;
}

interface Htg {
  name: string;
  symbol: string;
}

interface Aud {
  name: string;
  symbol: string;
}

interface Eur {
  name: string;
  symbol: string;
}

interface Bgn {
  name: string;
  symbol: string;
}

interface Cve {
  name: string;
  symbol: string;
}

interface Irr {
  name: string;
  symbol: string;
}

interface Pkr {
  name: string;
  symbol: string;
}

interface Ngn {
  name: string;
  symbol: string;
}

interface Inr {
  name: string;
  symbol: string;
}

interface Usd {
  name: string;
  symbol: string;
}

interface Try {
  name: string;
  symbol: string;
}

interface Tjs {
  name: string;
  symbol: string;
}

interface Mwk {
  name: string;
  symbol: string;
}

interface Afn {
  name: string;
  symbol: string;
}

interface Syp {
  name: string;
  symbol: string;
}

interface Krw {
  name: string;
  symbol: string;
}

interface Mvr {
  name: string;
  symbol: string;
}

interface Lbp {
  name: string;
  symbol: string;
}

interface Dop {
  name: string;
  symbol: string;
}

interface Bob {
  name: string;
  symbol: string;
}

interface Kgs {
  name: string;
  symbol: string;
}

interface Ggp {
  name: string;
  symbol: string;
}

interface Xcd {
  name: string;
  symbol: string;
}

interface Idr {
  name: string;
  symbol: string;
}

interface Pen {
  name: string;
  symbol: string;
}

interface Nzd {
  name: string;
  symbol: string;
}

interface All {
  name: string;
  symbol: string;
}

interface Fjd {
  name: string;
  symbol: string;
}

interface Top {
  name: string;
  symbol: string;
}

interface Bwp {
  name: string;
  symbol: string;
}

interface Vuv {
  name: string;
  symbol: string;
}

interface Jpy {
  name: string;
  symbol: string;
}

interface Ern {
  name: string;
  symbol: string;
}

interface Mxn {
  name: string;
  symbol: string;
}

interface Crc {
  name: string;
  symbol: string;
}

interface Ugx {
  name: string;
  symbol: string;
}

interface Czk {
  name: string;
  symbol: string;
}

interface Jep {
  name: string;
  symbol: string;
}

interface Xpf {
  name: string;
  symbol: string;
}

interface Kes {
  name: string;
  symbol: string;
}

interface Nio {
  name: string;
  symbol: string;
}

interface Uzs {
  name: string;
  symbol: string;
}

interface Scr {
  name: string;
  symbol: string;
}

interface Egp {
  name: string;
  symbol: string;
}

interface Ils {
  name: string;
  symbol: string;
}

interface Jod {
  name: string;
  symbol: string;
}

interface Lak {
  name: string;
  symbol: string;
}

interface Tmt {
  name: string;
  symbol: string;
}

interface Aed {
  name: string;
  symbol: string;
}

interface Xof {
  name: string;
  symbol: string;
}

interface Cny {
  name: string;
  symbol: string;
}

interface Bsd {
  name: string;
  symbol: string;
}

interface Lsl {
  name: string;
  symbol: string;
}

interface Zar {
  name: string;
  symbol: string;
}

interface Pgk {
  name: string;
  symbol: string;
}

interface Azn {
  name: string;
  symbol: string;
}

interface Tnd {
  name: string;
  symbol: string;
}

interface Sgd {
  name: string;
  symbol: string;
}

interface Ron2 {
  name: string;
  symbol: string;
}

interface Sos {
  name: string;
  symbol: string;
}
// ...
interface Currencies {
  TTD?: Ttd;
  SAR?: Sar;
  GYD?: Gyd;
  XAF?: Xaf;
  GBP?: Gbp;
  IMP?: Imp;
  SHP?: Shp;
  HTG?: Htg;
  AUD?: Aud;
  EUR?: Eur;
  BGN?: Bgn;
  CVE?: Cve;
  IRR?: Irr;
  PKR?: Pkr;
  NGN?: Ngn;
  INR?: Inr;
  USD?: Usd;
  TRY?: Try;
  TJS?: Tjs;
  MWK?: Mwk;
  AFN?: Afn;
  SYP?: Syp;
  KRW?: Krw;
  MVR?: Mvr;
  LBP?: Lbp;
  DOP?: Dop;
  BOB?: Bob;
  KGS?: Kgs;
  GGP?: Ggp;
  XCD?: Xcd;
  IDR?: Idr;
  PEN?: Pen;
  NZD?: Nzd;
  ALL?: All;
  FJD?: Fjd;
  TOP?: Top;
  BWP?: Bwp;
  VUV?: Vuv;
  JPY?: Jpy;
  ERN?: Ern;
  MXN?: Mxn;
  CRC?: Crc;
  UGX?: Ugx;
  CZK?: Czk;
  JEP?: Jep;
  XPF?: Xpf;
  KES?: Kes;
  NIO?: Nio;
  UZS?: Uzs;
  SCR?: Scr;
  EGP?: Egp;
  ILS?: Ils;
  JOD?: Jod;
  LAK?: Lak;
  TMT?: Tmt;
  AED?: Aed;
  XOF?: Xof;
  CNY?: Cny;
  BSD?: Bsd;
  LSL?: Lsl;
  ZAR?: Zar;
  PGK?: Pgk;
  AZN?: Azn;
  TND?: Tnd;
  SGD?: Sgd;
  RON?: Ron2;
  SOS?: Sos;
  LRD?: Lrd;
  KID?: Kid;
  NPR?: Npr;
  VND?: Vnd;
  RUB?: Rub;
  DZD?: Dzd;
  TVD?: Tvd;
  CHF?: Chf;
  MGA?: Mga;
  MYR?: Myr;
  GIP?: Gip;
  JMD?: Jmd;
  BDT?: Bdt;
  KWD?: Kwd;
  NOK?: Nok;
  KMF?: Kmf;
  BAM?: Bam;
  ETB?: Etb;
  ANG?: Ang;
  SDG?: Sdg;
  MAD?: Mad;
  MRU?: Mru;
  MNT?: Mnt;
  SLL?: Sll;
  BND?: Bnd;
  PHP?: Php;
  CDF?: Cdf;
  STN?: Stn;
  FKP?: Fkp;
  MMK?: Mmk;
  QAR?: Qar;
  NAD?: Nad;
  SSP?: Ssp;
  GNF?: Gnf;
  SRD?: Srd;
  ARS?: Ars;
  CAD?: Cad;
  BYN?: Byn;
  BZD?: Bzd;
  ZWL?: Zwl;
  GMD?: Gmd;
  GEL?: Gel;
  TZS?: Tzs;
  MUR?: Mur;
  PAB?: Pab;
  AOA?: Aoa;
  SEK?: Sek;
  RSD?: Rsd;
  LYD?: Lyd;
  BBD?: Bbd;
  COP?: Cop;
  ZMW?: Zmw;
  HNL?: Hnl;
  AMD?: Amd;
  THB?: Thb;
  OMR?: Omr;
  BRL?: Brl;
  SBD?: Sbd;
  PLN?: Pln;
  MKD?: Mkd2;
  IQD?: Iqd;
  AWG?: Awg;
  DKK?: Dkk;
  FOK?: Fok;
  BHD?: Bhd;
  KPW?: Kpw;
  KYD?: Kyd;
  BTN?: Btn;
  MZN?: Mzn;
  GHS?: Ghs;
  GTQ?: Gtq;
  LKR?: Lkr;
  UYU?: Uyu;
  CLP?: Clp;
  MDL?: Mdl;
  HUF?: Huf;
  SZL?: Szl;
  PYG?: Pyg;
  ISK?: Isk;
  CKD?: Ckd;
  CUC?: Cuc;
  CUP?: Cup;
  VES?: Ves;
  WST?: Wst;
  KHR?: Khr;
  HKD?: Hkd;
  YER?: Yer;
  DJF?: Djf;
  HRK?: Hrk;
  KZT?: Kzt;
  BMD?: Bmd;
  RWF?: Rwf;
  MOP?: Mop;
  BIF?: Bif;
  UAH?: Uah;
  TWD?: Twd;
}

interface Idd {
  root?: string;
  suffixes?: string[];
}

interface Languages {
  eng?: string;
  ara?: string;
  fra?: string;
  glv?: string;
  hat?: string;
  est?: string;
  bul?: string;
  por?: string;
  fas?: string;
  lav?: string;
  urd?: string;
  hin?: string;
  tam?: string;
  spa?: string;
  tur?: string;
  rus?: string;
  tgk?: string;
  nya?: string;
  prs?: string;
  pus?: string;
  tuk?: string;
  pih?: string;
  kor?: string;
  ita?: string;
  div?: string;
  aym?: string;
  grn?: string;
  que?: string;
  deu?: string;
  kir?: string;
  nfr?: string;
  ind?: string;
  niu?: string;
  sqi?: string;
  fij?: string;
  hif?: string;
  ton?: string;
  tsn?: string;
  pau?: string;
  bis?: string;
  srp?: string;
  slk?: string;
  jpn?: string;
  tir?: string;
  mri?: string;
  nzs?: string;
  swa?: string;
  ces?: string;
  nrf?: string;
  cha?: string;
  swe?: string;
  uzb?: string;
  crs?: string;
  nld?: string;
  lao?: string;
  ell?: string;
  zho?: string;
  sot?: string;
  hmo?: string;
  tpi?: string;
  aze?: string;
  msa?: string;
  mah?: string;
  ron?: string;
  som?: string;
  gil?: string;
  nep?: string;
  vie?: string;
  tvl?: string;
  gsw?: string;
  roh?: string;
  mlg?: string;
  cat?: string;
  jam?: string;
  ben?: string;
  nno?: string;
  nob?: string;
  smi?: string;
  pap?: string;
  zdj?: string;
  bos?: string;
  hrv?: string;
  amh?: string;
  bar?: string;
  fin?: string;
  kon?: string;
  lin?: string;
  tet?: string;
  ber?: string;
  mey?: string;
  mon?: string;
  fil?: string;
  lua?: string;
  lit?: string;
  nor?: string;
  gle?: string;
  mya?: string;
  afr?: string;
  her?: string;
  hgm?: string;
  kwn?: string;
  loz?: string;
  ndo?: string;
  bel?: string;
  lat?: string;
  bjz?: string;
  bwg?: string;
  kck?: string;
  khi?: string;
  ndc?: string;
  nde?: string;
  sna?: string;
  toi?: string;
  tso?: string;
  ven?: string;
  xho?: string;
  zib?: string;
  smo?: string;
  mlt?: string;
  kat?: string;
  mfe?: string;
  ltz?: string;
  hye?: string;
  sag?: string;
  tha?: string;
  slv?: string;
  pol?: string;
  mkd?: string;
  arc?: string;
  ckb?: string;
  dan?: string;
  fao?: string;
  dzo?: string;
  heb?: string;
  sin?: string;
  nau?: string;
  kal?: string;
  hun?: string;
  ssw?: string;
  isl?: string;
  pov?: string;
  rar?: string;
  cnr?: string;
  tkl?: string;
  khm?: string;
  kaz?: string;
  cal?: string;
  nbl?: string;
  nso?: string;
  zul?: string;
  kin?: string;
  run?: string;
  ukr?: string;
}
// ...
interface Translations {
  ara: Ara2;
  ces: Ces2;
  cym: Cym;
  deu: Deu2;
  est: Est2;
  fin: Fin2;
  fra: Fra2;
  hrv?: Hrv2;
  hun: Hun2;
  ita: Ita2;
  jpn?: Jpn2;
  kor: Kor2;
  nld: Nld2;
  per?: Per;
  pol: Pol2;
  por: Por2;
  rus: Rus2;
  slk: Slk2;
  spa: Spa2;
  swe: Swe2;
  urd: Urd2;
  zho?: Zho2;
}

interface Eng2 {
  f: string;
  m: string;
}

interface Fra3 {
  f: string;
  m: string;
}

interface Demonyms {
  eng: Eng2;
  fra?: Fra3;
}

interface Maps {
  googleMaps: string;
  openStreetMaps: string;
}

interface Gini {
  '1992'?: number;
  '1998'?: number;
  '2017'?: number;
  '2012'?: number;
  '2018'?: number;
  '2015'?: number;
  '2011'?: number;
  '2019'?: number;
  '2016'?: number;
  '2003'?: number;
  '2013'?: number;
  '2010'?: number;
  '2014'?: number;
  '2009'?: number;
  '2005'?: number;
  '2006'?: number;
  '2004'?: number;
  '1999'?: number;
  '2008'?: number;
}

interface Car {
  signs?: string[];
  side: string;
}

interface Flags {
  png: string;
  svg: string;
}

interface CoatOfArms {
  png?: string;
  svg?: string;
}

interface CapitalInfo {
  latlng?: number[];
}

interface PostalCode {
  format: string;
  regex?: string;
}

export interface ICountry {
  name: Name;
  tld?: string[];
  cca2: string;
  ccn3?: string;
  cca3: string;
  cioc?: string;
  independent?: boolean;
  status: string;
  unMember: boolean;
  currencies?: Currencies;
  idd: Idd;
  capital?: string[];
  altSpellings: string[];
  region: string;
  subregion?: string;
  languages?: Languages;
  translations: Translations;
  latlng: number[];
  landlocked: boolean;
  borders?: string[];
  area: number;
  demonyms?: Demonyms;
  flag?: string;
  maps: Maps;
  population: number;
  gini?: Gini;
  fifa?: string;
  car: Car;
  timezones: string[];
  continents: string[];
  flags: Flags;
  coatOfArms: CoatOfArms;
  startOfWeek: string;
  capitalInfo: CapitalInfo;
  postalCode?: PostalCode;
}